check if there's any other recoveried rtp packet

+++ modules/access/rtp/rtp.h
@@ -102,8 +102,8 @@ struct demux_sys_t
     vlc_thread_t  thread;
 
     bool b_intfec;
-    uint8_t max_rtpdepth;
-    uint8_t max_fecdepth;
+    uint16_t max_rtpdepth;
+    uint16_t max_fecdepth;
 
     mtime_t       timeout;
     uint16_t      max_dropout; /**< Max packet forward misordering */
